Change package of navigator and view model
