--- features/ranking/src/main/kotlin/com/glacierpower/tennisapp/features/ranking/adapter/navigation/RankingNavigator.kt
@@ -1,5 +1,0 @@
-package com.glacierpower.tennisapp.features.ranking.adapter.navigation
-
-interface RankingNavigator {
-    fun navigateToProfile(id: String)
-}
