+++ features/ranking/src/main/kotlin/com/glacierpower/tennisapp/features/ranking/navigation/RankingNavigator.kt
@@ -1,0 +1,5 @@
+package com.glacierpower.tennisapp.features.ranking.navigation
+
+interface RankingNavigator {
+    fun navigateToProfile(id: String)
+}
